pyproject.toml
[project]
name = "papers-vibe-reading"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
]
